Fix Phase 5: eliminate all remaining compiler warnings in src/enzo

+++ src/enzo/Grid_CorrectForRefinedFluxes.C
@@ -448,7 +448,7 @@ int grid::CorrectForRefinedFluxes(fluxes *InitialFluxes,
 	     total number density summed over ionization, etc.) */
 	
 	  if (Coordinate == Cartesian) {
-        for (int index = 0; index < fieldNumberList.size(); ++index) {
+        for (size_t index = 0; index < fieldNumberList.size(); ++index) {
           field = fieldNumberList[index];
           for (k = Start[2]; k <= End[2]; k++) {
             for (j = Start[1]; j <= End[1]; j++) {
@@ -564,7 +564,7 @@ int grid::CorrectForRefinedFluxes(fluxes *InitialFluxes,
 				   RefinedFluxes->LeftFluxes[field][dim][FluxIndex],
 				   CorrectionAmountLeft,
 				   i, j, k, dim, field);
-			  for (int undoIndex = 0; undoIndex < fieldNumberList.size(); ++undoIndex) {
+			  for (size_t undoIndex = 0; undoIndex < fieldNumberList.size(); ++undoIndex) {
                 ffield = fieldNumberList[undoIndex];
                 // If the flux correction was already applied, then undo it.
                 if (undoIndex <= index) {
@@ -586,7 +586,7 @@ int grid::CorrectForRefinedFluxes(fluxes *InitialFluxes,
 				   RefinedFluxes->RightFluxes[field][dim][RefinedFluxIndex],
 				   CorrectionAmountLeft,
 				   i, j, k, dim, field);
-			  for (int undoIndex = 0; undoIndex < fieldNumberList.size(); ++undoIndex) {
+			  for (size_t undoIndex = 0; undoIndex < fieldNumberList.size(); ++undoIndex) {
                 ffield = fieldNumberList[undoIndex];
                 // If the flux correction was already applied, then undo it.
                 if (undoIndex <= index) {
@@ -614,7 +614,7 @@ int grid::CorrectForRefinedFluxes(fluxes *InitialFluxes,
 				   RefinedFluxes->RightFluxes[field][dim][FluxIndex],
 				   CorrectionAmountRight,
 				   i, j, k, dim, field);
-			  for (int undoIndex = 0; undoIndex < fieldNumberList.size(); ++undoIndex) {
+			  for (size_t undoIndex = 0; undoIndex < fieldNumberList.size(); ++undoIndex) {
                 ffield = fieldNumberList[undoIndex];
                 // If the flux correction was already applied, then undo it.
                 if (undoIndex <= index) {
@@ -636,7 +636,7 @@ int grid::CorrectForRefinedFluxes(fluxes *InitialFluxes,
 				   RefinedFluxes->RightFluxes[field][dim][RefinedFluxIndex],
 				   CorrectionAmountRight,
 				   i, j, k, dim, field);
-			  for (int undoIndex = 0; undoIndex < fieldNumberList.size(); ++undoIndex) {
+			  for (size_t undoIndex = 0; undoIndex < fieldNumberList.size(); ++undoIndex) {
                 ffield = fieldNumberList[undoIndex];
                 // If the flux correction was already applied, then undo it.
                 if (undoIndex <= index) {
@@ -655,13 +655,13 @@ int grid::CorrectForRefinedFluxes(fluxes *InitialFluxes,
 		  }// for (i = Start[0]; i <= End[0]; i++) 
 		} // for (j = Start[1]; j <= End[1]; j++)
 	      } // for (k = Start[2]; k <= End[2]; k++)
-	    } // for (int index = 0; index < fieldNumberList.size(); ++index)
+	    } // for (size_t index = 0; index < fieldNumberList.size(); ++index)
 	  }
 
 
       if (Coordinate == Cylindrical) {
         FLOAT xr, xl, xc, geofacr, geofacl;
-        for (int index = 0; index < fieldNumberList.size(); ++index) {
+        for (size_t index = 0; index < fieldNumberList.size(); ++index) {
           field = fieldNumberList[index];
 	  for (k = Start[2]; k <= End[2]; k++) {
 	    for (j = Start[1]; j <= End[1]; j++) {
@@ -732,7 +732,7 @@ int grid::CorrectForRefinedFluxes(fluxes *InitialFluxes,
 			   i, j, k, dim, field);
 			   }*/
 		  /* If new density & energy is < 0 then undo the flux correction. */
-			  for (int undoIndex = 0; undoIndex < fieldNumberList.size(); ++undoIndex) {
+			  for (size_t undoIndex = 0; undoIndex < fieldNumberList.size(); ++undoIndex) {
                 ffield = fieldNumberList[undoIndex];
             if (undoIndex <= index) {
               BaryonField[ffield][FieldIndex] -= (
@@ -759,7 +759,7 @@ int grid::CorrectForRefinedFluxes(fluxes *InitialFluxes,
 			   i, j, k, dim, Offset, field);
 			   }*/
 		  /* If new density & energy is < 0 then undo the flux correction. */
-			  for (int undoIndex = 0; undoIndex < fieldNumberList.size(); ++undoIndex) {
+			  for (size_t undoIndex = 0; undoIndex < fieldNumberList.size(); ++undoIndex) {
                 ffield = fieldNumberList[undoIndex];
             if (undoIndex <= index) {
               BaryonField[ffield][FieldIndex + Offset] += (
